+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "express": "^4.10.7",
     "hbs": "^2.8.0",
+    "passport": "^0.2.1",
     "showdown": "^0.3.1",
     "socket.io": "^1.2.1",
     "sqlite3": "^3.0.4"
